requirements.txt
Flask==3.0.2
flask_cors==4.0.1
pillow==10.3.0
tensorflow==2.16.1